+++ libcruft/gethostbyname2_r.c
@@ -22,7 +22,7 @@ extern int __dns_gethostbyx_r(const char* name, struct hostent* result,
 int gethostbyname2_r(const char* name, int AF, struct hostent* result,
 				char *buf, size_t buflen,
 				struct hostent **RESULT, int *h_errnop) {
-  int L=strlen(name);
+  size_t L=strlen(name);
   int lookfor=0;
   switch (AF) {
   case AF_INET: lookfor=1; break;
